ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
qdrant-client==1.12.0
langchain==0.3.0
langchain-community==0.3.0
python-dotenv==1.0.1
pydantic-settings==2.5.0
